# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aiexec",
    "version": "1.4.2"
  },
  "paths": {
...
    "/api/v1/monitor/builds": {
      "get": {
        "tags": ["Monitor"],
        "summary": "Get Vertex Builds",
        "operationId": "get_vertex_builds_api_v1_monitor_builds_get",
        "parameters": [
          {
            "name": "flow_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VertexBuildMapModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Monitor"],
        "summary": "Delete Vertex Builds",
        "operationId": "delete_vertex_builds_api_v1_monitor_builds_delete",
        "parameters": [
          {
            "name": "flow_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flow Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitor/messages": {
      "get": {
        "tags": ["Monitor"],
        "summary": "Get Messages",
        "operationId": "get_messages_api_v1_monitor_messages_get",
        "parameters": [
          {
            "name": "flow_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flow Id"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "sender",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sender"
            }
          },
          {
            "name": "sender_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sender Name"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "timestamp",
              "title": "Order By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Get Messages Api V1 Monitor Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Monitor"],
        "summary": "Delete Messages",
        "operationId": "delete_messages_api_v1_monitor_messages_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "title": "Message Ids"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitor/messages/{message_id}": {
      "put": {
        "tags": ["Monitor"],
        "summary": "Update Message",
        "operationId": "update_message_api_v1_monitor_messages__message_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitor/messages/session/{old_session_id}": {
      "patch": {
        "tags": ["Monitor"],
        "summary": "Update Session Id",
        "operationId": "update_session_id_api_v1_monitor_messages_session__old_session_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "parameters": [
          {
            "name": "old_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Old Session Id"
            }
          },
          {
            "name": "new_session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The new session ID to update to",
              "title": "New Session Id"
            },
            "description": "The new session ID to update to"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Update Session Id Api V1 Monitor Messages Session  Old Session Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitor/messages/session/{session_id}": {
      "delete": {
        "tags": ["Monitor"],
        "summary": "Delete Messages Session",
        "operationId": "delete_messages_session_api_v1_monitor_messages_session__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitor/transactions": {
      "get": {
        "tags": ["Monitor"],
        "summary": "Get Transactions",
        "operationId": "get_transactions_api_v1_monitor_transactions_get",
        "parameters": [
          {
            "name": "flow_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flow Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Size"
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TransactionTable_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentBlock": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "contents": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Contents"
          },
          "allow_markdown": {
            "type": "boolean",
            "title": "Allow Markdown",
            "default": true
          },
          "media_url": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Url"
          }
        },
        "type": "object",
        "required": ["title", "contents"],
        "title": "ContentBlock",
        "description": "A block of content that can contain different types of content."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageRead": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "sender_name": {
            "type": "string",
            "title": "Sender Name"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files"
          },
          "error": {
            "type": "boolean",
            "title": "Error",
            "default": false
          },
          "edit": {
            "type": "boolean",
            "title": "Edit",
            "default": false
          },
          "properties": {
            "$ref": "#/components/schemas/Properties"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "message"
          },
          "content_blocks": {
            "items": {
              "$ref": "#/components/schemas/ContentBlock"
            },
            "type": "array",
            "title": "Content Blocks"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "flow_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flow Id"
          }
        },
        "type": "object",
        "required": [
          "sender",
          "sender_name",
          "session_id",
          "text",
          "id",
          "flow_id"
        ],
        "title": "MessageRead"
      },
      "MessageResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "flow_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flow Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "sender_name": {
            "type": "string",
            "title": "Sender Name"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files",
            "default": []
          },
          "edit": {
            "type": "boolean",
            "title": "Edit"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "properties": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Properties"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "content_blocks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ContentBlock"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Blocks"
          }
        },
        "type": "object",
        "required": ["sender", "sender_name", "session_id", "text", "edit"],
        "title": "MessageResponse"
      },
      "MessageUpdate": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "sender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender"
          },
          "sender_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Name"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "edit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edit"
          },
          "error": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "properties": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Properties"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "MessageUpdate"
      },
...
      "Page_TransactionTable_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionTable"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          }
        },
        "type": "object",
        "required": ["items", "page", "size"],
        "title": "Page[TransactionTable]"
      },
...
      "Properties": {
        "properties": {
          "text_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Color"
          },
          "background_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background Color"
          },
          "edited": {
            "type": "boolean",
            "title": "Edited",
            "default": false
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "allow_markdown": {
            "type": "boolean",
            "title": "Allow Markdown",
            "default": false
          },
          "positive_feedback": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Positive Feedback"
          },
          "state": {
            "type": "string",
            "enum": ["partial", "complete"],
            "title": "State",
            "default": "complete"
          },
          "targets": {
            "items": {},
            "type": "array",
            "title": "Targets",
            "default": []
          }
        },
        "type": "object",
        "title": "Properties"
      },
...
      "Source": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "The id of the source component."
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "The display name of the source component."
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "The source of the message. Normally used to display the model name (e.g. 'gpt-4o')"
          }
        },
        "type": "object",
        "title": "Source"
      },
...
      "TransactionTable": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "vertex_id": {
            "type": "string",
            "title": "Vertex Id"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "inputs": {
            "type": "object",
            "title": "Inputs"
          },
          "outputs": {
            "type": "object",
            "title": "Outputs"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "flow_id": {
            "type": "string",
            "format": "uuid",
            "title": "Flow Id"
          },
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": ["vertex_id", "status", "flow_id"],
        "title": "TransactionTable"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
      "VertexBuildMapModel": {
        "properties": {
          "vertex_builds": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/VertexBuildTable"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Vertex Builds"
          }
        },
        "type": "object",
        "required": ["vertex_builds"],
        "title": "VertexBuildMapModel"
      },
...
      "VertexBuildTable": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "data": {
            "type": "object",
            "title": "Data"
          },
          "artifacts": {
            "type": "object",
            "title": "Artifacts"
          },
          "params": {
            "type": "string",
            "title": "Params"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "flow_id": {
            "type": "string",
            "format": "uuid",
            "title": "Flow Id"
          },
          "build_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Id"
          }
        },
        "type": "object",
        "required": ["id", "valid", "flow_id"],
        "title": "VertexBuildTable"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/login"
          }
        }
      },
      "API key query": {
        "type": "apiKey",
        "in": "query",
        "name": "x-api-key"
      },
      "API key header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}